Moved F-API helper functions from data_offload.py to parallel/field_api.py
