+++ loki/transformations/parallel/field_api.py
@@ -9,16 +9,19 @@
 Transformation utilities to manage and inject FIELD-API boilerplate code.
 """
 
+from enum import Enum
 from loki.expression import symbols as sym
 from loki.ir import (
     nodes as ir, FindNodes, FindVariables, Transformer
 )
+from loki.scope import Scope
 from loki.logging import warning
 from loki.tools import as_tuple
 
-
 __all__ = [
-    'remove_field_api_view_updates', 'add_field_api_view_updates'
+    'remove_field_api_view_updates', 'add_field_api_view_updates', 'get_field_type',
+    'field_new', 'field_delete', 'field_get_device_data', 'field_sync_host',
+    'FieldAPITransferType'
 ]
 
 
@@ -26,9 +29,7 @@ def remove_field_api_view_updates(routine, field_group_types, dim_object=None):
     """
     Remove FIELD API boilerplate calls for view updates of derived types.
 
-    This utility is intended to remove the IFS-specific group type
-    objects that provide block-scope view pointers to deep kernel
-    trees. It will remove all calls to ``UPDATE_VIEW`` on derive-type
+    This utility is intended to remove the IFS-specific group type objects that provide block-scope view pointers to deep kernel trees. It will remove all calls to ``UPDATE_VIEW`` on derive-type
     objects with the respective types.
 
     Parameters
@@ -150,3 +151,66 @@ def add_field_api_view_updates(routine, dimension, field_group_types, dim_object
             return loop
 
     routine.body = InsertFieldAPIViewsTransformer().visit(routine.body, scope=routine)
+
+
+def get_field_type(a: sym.Array) -> sym.DerivedType:
+    """
+    Returns the corresponding FIELD API type for an array.
+
+    This transformation is IFS specific and assumes that the
+    type is an array declared with one of the IFS type specifiers, e.g. KIND=JPRB
+    """
+    type_map = ["jprb",
+                "jpit",
+                "jpis",
+                "jpim",
+                "jpib",
+                "jpia",
+                "jprt",
+                "jprs",
+                "jprm",
+                "jprd",
+                "jplm"]
+
+    type_name = a.type.kind.name
+    assert type_name.lower() in type_map, ('Error array type kind is: '
+                                           f'"{type_name}" which is not a valid IFS type specifier')
+    rank = len(a.shape)
+    field_type = sym.DerivedType(name="field_" + str(rank) + type_name[2:4])
+    return field_type
+
+
+def field_new(field_ptr, data, scope):
+    return ir.CallStatement(name=sym.ProcedureSymbol('FIELD_NEW', scope=scope),
+                            arguments=(field_ptr,), kwarguments=(('DATA', data),))
+
+
+def field_delete(field_ptr, scope):
+    return ir.CallStatement(name=sym.ProcedureSymbol('FIELD_DELETE', scope=scope),
+                            arguments=(field_ptr,))
+
+
+class FieldAPITransferType(Enum):
+    READ_ONLY = 1
+    READ_WRITE = 2
+    WRITE_ONLY = 3
+
+
+def field_get_device_data(field_ptr, dev_ptr, transfer_type: FieldAPITransferType, scope: Scope):
+    assert isinstance(transfer_type, FieldAPITransferType)
+    if transfer_type == FieldAPITransferType.READ_ONLY:
+        suffix = 'RDONLY'
+    if transfer_type == FieldAPITransferType.READ_WRITE:
+        suffix = 'RDWR'
+    if transfer_type == FieldAPITransferType.WRITE_ONLY:
+        suffix = 'WRONLY'
+    procedure_name = 'GET_DEVICE_DATA_' + suffix
+    return ir.CallStatement(name=sym.ProcedureSymbol(procedure_name, parent=field_ptr, scope=scope),
+                            arguments=(dev_ptr.clone(dimensions=None),), )
+
+
+def field_sync_host(field_ptr, scope):
+    procedure_name = 'SYNC_HOST_RDWR'
+    return ir.CallStatement(name=sym.ProcedureSymbol(procedure_name, parent=field_ptr, scope=scope), arguments=())
+
+
